feat: implement context compaction service and integrate with Orchestrator
