+++ crates/forge_domain/src/lib.rs
@@ -4,6 +4,7 @@ mod attachment;
 mod chat_request;
 mod chat_response;
 mod compact;
+mod compaction;
 mod compaction_result;
 mod conversation_html;
 mod update;
@@ -49,6 +50,7 @@ pub use attachment::*;
 pub use chat_request::*;
 pub use chat_response::*;
 pub use compact::*;
+pub use compaction::*;
 pub use compaction_result::*;
 pub use context::*;
 pub use conversation::*;
